Use size_t for lengths.

+++ src/b64/b64.h
@@ -10,6 +10,6 @@
 /* #define DEBUG_B64 */
 
 void b64_encode( uint8_t* indata, long indata_len, bstring outstring, int linesz );
-uint8_t* b64_decode( long* outdata_len, bstring instring );
+uint8_t* b64_decode( size_t* outdata_len, bstring instring );
 
 #endif /* B64_H */
